metrics logging, error-handling, serialize model
this diff does a bunch of stuff:
- fixes the `env_dep_hyperparameter` logic. there was a bug in
  `algorithm_space.py` that overrides the `env_dep_hyperparameter`
  dictionary with the `hyperparameters` supplied to the create_ml_framework
  function
- implemented a `save` and `load` method for the cash controller
- fix bugs in the hyperparameter configuration of decision trees,
  kernal_pca, nystroem_sampler.
- added error handling module that explicitly documents the errors
  and warnings that are "valid", that is, those that justifiably
  lead to a negative reward signal.
- fixed bug in error reward signal, which always assigns
  `self._error_reward = -1`.
- created additional metafeatures: dataset name, number of features.
- add environment variables to `example_rnn_cash_controller_experiment.py`

+++ deep_cash/examples/example_rnn_cash_controller.py
@@ -11,28 +11,41 @@ import torch
 from pathlib import Path
 from shutil import rmtree
 from sklearn.externals import joblib
-from sklearn.metrics import roc_auc_score, f1_score
-
-from deep_cash.cash_controller import CASHController
-from deep_cash.cash_reinforce import CASHReinforce
+from sklearn.metrics import f1_score
 
 from deep_cash.algorithm_space import AlgorithmSpace
 from deep_cash.task_environment import TaskEnvironment
+from deep_cash.cash_controller import CASHController
+from deep_cash.cash_reinforce import CASHReinforce
+from deep_cash.loggers import get_loggers
+from deep_cash.utils import METAFEATURE_DIM
 
 
-data_path = Path(os.path.dirname(__file__)) / "artifacts"
+data_path = Path(os.environ.get(
+    "DEEP_CASH_OUT_PATH", os.path.dirname(__file__) + "/artifacts"))
 
 # hyperparameters
-metafeatures = ["number_of_examples"]
-learning_rate = 0.005
+n_episodes = int(os.environ.get("DEEP_CASH_N_EPISODES", 500))
+n_iter = int(os.environ.get("DEEP_CASH_N_ITER", 50))
+learning_rate = float(os.environ.get("DEEP_CASH_LEARNING_RATE", 0.005))
+error_reward = int(os.environ.get("DEEP_CASH_ERROR_REWARD", -1))
+logger_name = os.environ.get("DEEP_CASH_LOGGER", None)
+logger = get_loggers().get(os.environ.get("DEEP_CASH_LOGGER", None), None)
+fit_verbose = int(os.environ.get("DEEP_CASH_FIT_VERBOSE", 1))
+
 hidden_size = 30
 output_size = 30
 n_layers = 3
 
 
 t_env = TaskEnvironment(
-    f1_score, scorer_kwargs={"average": "weighted"}, random_state=100,
-    per_framework_time_limit=5)
+    f1_score,
+    scorer_kwargs={"average": "weighted"},
+    random_state=100,
+    per_framework_time_limit=60,
+    per_framework_memory_limit=3077,
+    error_reward=error_reward,
+    reward_transformer=lambda x: x)
 
 # create algorithm space
 a_space = AlgorithmSpace(
@@ -41,7 +54,7 @@ a_space = AlgorithmSpace(
     hyperparam_with_none_token=False)
 
 controller = CASHController(
-    metafeature_size=len(metafeatures),
+    metafeature_size=METAFEATURE_DIM,
     input_size=a_space.n_components,
     hidden_size=hidden_size,
     output_size=output_size,
@@ -51,8 +64,8 @@ controller = CASHController(
     dropout_rate=0.2,
     num_rnn_layers=n_layers)
 
-reinforce = CASHReinforce(controller, t_env)
-reinforce.fit(n_episodes=500, n_iter=50)
+reinforce = CASHReinforce(controller, t_env, metrics_logger=logger)
+reinforce.fit(n_episodes=n_episodes, n_iter=n_iter, verbose=fit_verbose)
 
 
 history = pd.DataFrame(reinforce.history())
@@ -69,7 +82,10 @@ history[[
 ]].to_csv(str(data_path / "rnn_cash_controller_experiment.csv"), index=False)
 
 mlf_path = data_path / "rnn_cash_controller_experiment_mlfs"
-rmtree(mlf_path)
+if mlf_path.exists():
+    rmtree(mlf_path)
 mlf_path.mkdir()
 for i, mlf in enumerate(history.best_mlfs):
     joblib.dump(mlf, mlf_path / ("best_mlf_episode_%d.pkl" % (i + 1)))
+
+controller.save(data_path / "rnn_cash_controller_experiment.pt")
